Database.RethinkDB.Typed

The expression type

newtype Expr a

A ReQL expression resulting in type a.

Constructors

Expr

Fields

unExpr :: ReQL

expr :: ToDatum a => a -> Expr (DatumOf a)

Execution

type family ResultOf a

run :: forall a. Result (ResultOf a) => RethinkDBHandle -> Expr a -> IO (ResultOf a)

Primitive types

type Datum = Datum

type Object = Object

type Number = Double

type Time = ZonedTime

type String = Text

Sequence types

class Sequence s

data Stream a

data Selection a

data Table a

data Array a

data SingleSelection a

type family StreamOrArray s

Most ReQL functions can take both sequences and arrays, producing a new array for arrays and a stream otherwise.

This type family returns the output sequence type for input sequence type s.

Equations

StreamOrArray Array = Array

StreamOrArray s = Stream

Manipulating databases

dbCreate :: Text -> Expr Object

dbDrop :: Database -> Expr Object

dbList :: Expr (Array String)

Manipulating tables

tableCreate :: Table -> Expr Object

tableDrop :: Table -> Expr Object

tableList :: Database -> Expr (Array String)

class Observable s

Sequence types that can be watched for changes.

changes :: Observable s => Expr (s a) -> Expr (Stream Object)

Writing data

insert :: Expr (Array Object) -> Table -> Expr Object

class Updatable s

Sequence types whose members can be updated.

update :: Updatable s => (Expr Object -> Expr Object) -> Expr (s Object) -> Expr Object

replace :: Updatable s => (Expr Object -> Expr Object) -> Expr (s Object) -> Expr Object

delete :: Updatable s => Expr (s Object) -> Expr Object

sync :: Expr (Table a) -> Expr Object

Selecting data

get :: Expr String -> Expr (Table Object) -> Expr (SingleSelection Object)

table :: Text -> Expr (Table Object)

type family Filtered s

The result sequence type returned by filter given then input sequence type.

Equations

Filtered Table = Selection

Filtered Selection = Selection

Filtered Array = Array

Filtered Stream = Stream

filter :: Sequence s => (Expr a -> Expr Bool) -> Expr (s a) -> Expr (Filtered s a)

innerJoin :: (Sequence s, Sequence s') => (Expr a -> Expr b -> Expr Bool) -> Expr (s a) -> Expr (s' b) -> Expr (StreamOrArray s (a, b))

outerJoin :: (Sequence s, Sequence s') => (Expr a -> Expr b -> Expr Bool) -> Expr (s a) -> Expr (s' b) -> Expr (StreamOrArray s (a, b))

left :: Expr (a, b) -> Expr a

right :: Expr (a, b) -> Expr b

zip :: Sequence s => Expr (s (Object, Object)) -> Expr (StreamOrArray s Object)

Transformations

map :: Sequence s => (Expr a -> Expr b) -> Expr (s a) -> Expr (StreamOrArray s b)

withFields :: Sequence s => [Expr String] -> Expr (s Object) -> Expr (StreamOrArray s Object)

concatMap :: (Sequence s, Sequence s') => (Expr a -> Expr (s' b)) -> Expr (s a) -> Expr (StreamOrArray s b)

skip :: Sequence s => Expr Number -> Expr (s a) -> Expr (StreamOrArray s a)

limit :: Sequence s => Expr Number -> Expr (s a) -> Expr (StreamOrArray s a)

isEmpty :: Sequence s => Expr (s a) -> Expr Bool

union :: (Sequence s, Sequence s') => Expr (s a) -> Expr (s' a) -> Expr (StreamOrArray s a)

sample :: Sequence s => Expr Number -> Expr (s a) -> Expr (StreamOrArray s a)

Aggregation

reduce :: Sequence s => (Expr a -> Expr a -> Expr a) -> Expr (s a) -> Expr a

count :: Sequence s => Expr (s a) -> Expr Number

sum :: Sequence s => Expr (s Number) -> Expr Number

avg :: Sequence s => Expr (s Number) -> Expr Number

min :: Sequence s => Expr (s a) -> Expr a

max :: Sequence s => Expr (s a) -> Expr a

distinct :: Sequence s => Expr (s a) -> Expr (StreamOrArray s a)

contains :: Sequence s => Expr a -> Expr (s a) -> Expr Bool

Document manipulation

type family Manip a

The result of applying pluck or without to an expression.

pluck :: [Expr String] -> Expr a -> Expr (Manip a)

without :: [Expr String] -> Expr a -> Expr (Manip a)

append :: Expr a -> Expr (Array a) -> Expr (Array a)

prepend :: Expr a -> Expr (Array a) -> Expr (Array a)

difference :: Expr (Array a) -> Expr (Array a) -> Expr (Array a)

setInsert :: Expr a -> Expr (Array a) -> Expr (Array a)

setUnion :: Expr (Array a) -> Expr (Array a) -> Expr (Array a)

setIntersection :: Expr (Array a) -> Expr (Array a) -> Expr (Array a)

setDifference :: Expr (Array a) -> Expr (Array a) -> Expr (Array a)

class IsDatum a

(!..) :: Expr (Array a) -> Expr Number -> Expr a

Index an Array.

class SingleOrObj a

(!) :: (SingleOrObj o, IsDatum a) => Expr o -> Expr Text -> Expr a

Index an object.

keys :: SingleOrObj o => Expr o -> Expr (Array String)

values :: SingleOrObj o => Expr o -> Expr (Array Datum)

Math and logic

mod :: Expr Number -> Expr Number -> Expr Number

random :: Expr Number

round :: Expr Number -> Expr Number

floor :: Expr Number -> Expr Number

ceil :: Expr Number -> Expr Number

Time and date

now :: Expr Time

data BuildTime

Constructors

BuildTime

Fields

yearPart :: Expr Number

monthPart :: Expr Number

dayPart :: Expr Number

hourPart :: Expr Number

minutePart :: Expr Number

secondPart :: Expr Number

timezonePart :: Expr String

time :: BuildTime -> Expr Time

Control structures

forEach :: Sequence s => (Expr a -> Expr write) -> Expr (s a) -> Expr Object

range :: Expr Number -> Expr (Stream Number)

Type assertions

number :: Expr Number -> Expr Number

string :: Expr String -> Expr String

bool :: Expr Bool -> Expr Bool